requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
netaddr==1.3.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.0
yookassa==3.5.0
